- Extract UnZipTask from AssetsCache
