+++ app/src/main/java/me/devsaki/hentoid/util/AssetsCache.java
@@ -10,20 +10,14 @@ import com.thin.downloadmanager.DownloadRequest;
 import com.thin.downloadmanager.DownloadStatusListenerV1;
 import com.thin.downloadmanager.ThinDownloadManager;
 
-import org.apache.commons.io.IOUtils;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.util.Enumeration;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
 
 import me.devsaki.hentoid.HentoidApp;
 
@@ -132,7 +126,7 @@ public class AssetsCache {
         LogHelper.d(TAG, "Extracting cache files.");
         String zipFile = file.getAbsolutePath();
         String destinationPath = cacheDir.getPath();
-        new UnZipTask().execute(zipFile, destinationPath);
+        new ZipUtil.UnZipTask().execute(zipFile, destinationPath);
     }
 
     private static class UpdateCheckTask extends AsyncTask<String, Void, Void> {
@@ -174,76 +168,6 @@ public class AssetsCache {
                     LogHelper.d(TAG, "Bundled cache is current.");
                     unpackBundle();
                 }
-            }
-        }
-    }
-
-    private static class UnZipTask extends AsyncTask<String, Void, Boolean> {
-
-        @Override
-        protected Boolean doInBackground(String... params) {
-            String filePath = params[0];
-            String destinationPath = params[1];
-
-            File archive = new File(filePath);
-            try {
-                ZipFile zipfile = new ZipFile(archive);
-                for (Enumeration e = zipfile.entries(); e.hasMoreElements(); ) {
-                    ZipEntry entry = (ZipEntry) e.nextElement();
-                    unzipEntry(zipfile, entry, destinationPath);
-                }
-                zipfile.close();
-            } catch (Exception e) {
-                LogHelper.e(TAG, "Error while extracting file: " + archive, e);
-                return false;
-            }
-
-            return true;
-        }
-
-        @Override
-        protected void onPostExecute(Boolean aBoolean) {
-            LogHelper.d(TAG, "All files extracted without error: " + aBoolean);
-        }
-
-        private void unzipEntry(ZipFile zipfile, ZipEntry entry,
-                                String outputDir) throws IOException {
-
-            if (entry.isDirectory()) {
-                createDir(new File(outputDir, entry.getName()));
-                return;
-            }
-
-            File outputFile = new File(outputDir, entry.getName());
-            if (!outputFile.getParentFile().exists()) {
-                createDir(outputFile.getParentFile());
-            }
-
-            BufferedInputStream inputStream = new BufferedInputStream(
-                    zipfile.getInputStream(entry));
-            BufferedOutputStream outputStream = new BufferedOutputStream(
-                    new FileOutputStream(outputFile));
-
-            //noinspection TryFinallyCanBeTryWithResources
-            try {
-                IOUtils.copy(inputStream, outputStream);
-            } finally {
-                try {
-                    outputStream.close();
-                    inputStream.close();
-                } catch (IOException e) {
-                    // Ignore
-                }
-            }
-        }
-
-        private void createDir(File dir) {
-            if (dir.exists()) {
-                return;
-            }
-            LogHelper.d(TAG, "Creating dir: " + dir.getName());
-            if (!dir.mkdirs()) {
-                LogHelper.w(TAG, "Could not create dir: " + dir);
             }
         }
     }
